requirements.txt
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
torch
fastapi
uvicorn
pymupdf

